V5/project/requirements.txt
pandas==1.4.2
psycopg2-binary==2.9.3
PyYAML==6.0
category-encoders==2.4.1
scikit-learn==1.1.1
boto3==1.20.11
